Remove WebDavStore.createUri/decodeUri
