+++ mail/protocols/webdav/src/main/java/com/fsck/k9/mail/store/webdav/WebDavStoreUriCreator.java
@@ -17,7 +17,6 @@ public class WebDavStoreUriCreator {
      * @param server The {@link ServerSettings} object that holds the server settings.
      * @return A WebDavStore URI that holds the same information as the {@code server} parameter.
      * @see StoreConfig#getStoreUri()
-     * @see WebDavStore#decodeUri(String)
      */
     public static String create(ServerSettings server) {
         String userEnc = encodeUtf8(server.username);
